requirements.txt
Django==1.6.2
django-taggit==0.11.2
ipython==1.2.1
